+++ qrust-domain/src/main/java/com/qrust/common/infrastructure/FeignConfig.java
@@ -1,4 +1,4 @@
-package com.qrust.common.config;
+package com.qrust.common.infrastructure;
 
 import com.qrust.external.google.infrastructure.GoogleSafeBrowsingFeignClient;
 import org.springframework.cloud.openfeign.EnableFeignClients;
